Add FVP Cortex-A32 support
The Cortex-A32 model is very similar to the AEMv8 model and uses a
subset of commandline parameters to launch it.

The main difference is that the Cortex-A32 model only has 1 cluster and
therefore doesn't uspport the "-C cluster" options.

The model is Aarch32 mode only, but it uses the same DTB as the 64-bit
AEMv8 model.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>