sgi: reorganize sgi platform model-scripts
The existing structure of the sgi575 platform model-scripts is not
extensible when adding support for additional sgi platforms and
test cases. So the sgi platform model-scripts have been reorganized
as below

    model-scripts
         |
          ---sgi
              |
              |--- distro-boot.sh
              |--- busybox-boot.sh
               --- platforms
                       |
                        --- sgi575
                              |
                               --- run_model.sh

The new 'sgi' directory will hold all the test scripts. The new
'platforms' directory will hold the platform specific run_model.sh
script and there is a directory for each supported platform.

The distro-boot.sh is now moved to 'sgi' directory and the changes
required to align with this new structure has been included in this
script. There is also a new 'busybox-boot.sh' script as well to be
used for busybox boot tests.

Change-Id: I78445bf59f0690900f60a819500ab455e1c14179
Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>