sgi/distro_boot: fix useability issues
Modifiy the distro-boot script to allow it to accept sata disk images which
are placed at any location in the host's filesystem. Also, as part of the
restructuring of the model scripts, modify the distro-boot script to adapt
it to the changes in the location of the run-model.sh script.

Change-Id: I8401a6e5a7a3bf1e24106fd4b88b490ddc2068c8
Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>