sgi/common: rename check_boot_complete to parse_log_file
The 'check_boot_complete' function is renamed as 'parse_log_file'
and modified to make it usable by multiple test scripts. It now
accepts three parameters - log file name, search string and
timeout value.

It waits until the search string is found in the log file or the
timeout occurs. It returns '0' on success and '-1' on failure.

Change-Id: I90956c109bfd028606490781744a21009a1b8bdb
Signed-off-by: Lokesh B V <lokesh.bv@arm.com>