sgi/acs: allow prebuilt luv-live image to be specified
Add support for a command line option to specify a absolute path of the
luv-live image to be tested with. This is useful when executing ACS tests
with a pre-built version of the luv-live ACS image. To specify the
pre-built image, use the command line option "-v" and specify the
absolute path of the luv-live image as its parameter.

Change-Id: Ifaba3e7dc2571453e1f0647f1f657a0bfca2a33e
Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>