run_model.sh: enable unbuffered_output for scp and mcp consoles
Set unbuffered_output to true for SCP and MCP uart consoles. This will
write to the log file immediately instead of waiting for the model to
quit. This solves the partial scp and mcp logs issue when -j (AUTOMATE)
is set to true where the model is killed through SIGTERM signal when the
test is completed.

Change-Id: I8f6ac039bdc1393b494afd2d0f6d990c9935787c
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>