platforms/rdn2: use pcie_rc behind io macro for passing ahci sata image
RD-N2 FVP has been update to use IO Macro for the PCIe root complex. The
initial support has been added such that one IO Macro is enabled out of
the five IO Macros. Use the PCIe topology behind this IO Macro for
passing the achi sata disk image. Also for the root ports which are not
used by the software, pass the minimal PCIe hierarchy file to speed up
the launch of FVP.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Idf0e826086d1d04c3eec01112a504ff9111cd417