rdn2/run_model: do not create minimal pcie hierarchies
Prior to FVP version 11.15, FVP will create default pcie hierarchy if
a json hierarchy file is not passed to pciex* root complex. Creating
default hierarchy on all the ports slowed down the launch of the model.
To avoid this, a minimal end point topology file is passed to speed up
the launch of the model.

With 11.15, FVP will not create any pcie hierarchy if a topology file is
not passed to pciex*. Hence passing the minimal end point topology file
is not required anymore.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Ieadaa85c0d5a56e3937fa1db3ba97a0c9f4f4e21