pcie_hierarchies: remove minimal end point json file
Minimal endpoint json file was passed to the pcie root complex on which
the software enumeration will not be done. With 11.15 FVP, if the json
file is not passed to a pcie root complex, the pcie hierarchy will not
be created. So minimal endpoint json file is not used anymore. This
patch removes the stale file.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I7fb25fc624dc7fc7cb466b3f20baed28c99b9430