rdn2cfg1/run_model: do not create minimal pcie hierarchies
Prior to FVP version 11.15, FVP will create default pcie hierarchy if
a json hierarchy file is not passed to pciex* root complex. Creating
default hierarchy on all the ports slowed down the launch of the model.
To avoid this, a minimal end point topology file is passed to speed up
the launch of the model.

With 11.15, FVP will not create any pcie hierarchy if a topology file is
not passed to pciex*. Hence passing the minimal end point topology file
is not required anymore.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Id54c8b805286ee0daa585b91bc2d1c91f2af4a82