rdn2cfg2/run_model: update TZC_BYPASS_PARAM
Until 11.15 version, RD-N2-Cfg2 FVP had `mem` component which
instantiated `n` number of TZC's statistically and controlled the access
to number of regions required for a particular TZC during runtime.
Starting from 11.16 version, the `mem` component is removed and the
number of regions parameter (`tzcN.num_region`) for a particular TZC has
been removed as a model parameter and is configured statically by the
model.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I7def859a4ff1df1dc83f7a17e07086aa823737e4