rdn2cfg2/run_models: bypass trustzone controller (TZC) at lauch
FVP by default blocks all the transaction if trustzone controller (TZC)
is not configured by software. This is an opposite behaviour to hardware
which allows all access if the TZC is not configured by software. To
align with hardware behaviour, use model parameters to configure TZC to
be bypassed and allow all access until software reconfigures the TZC.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I4f5c22bce7fb1f6067c394db21b5329a439dd666