sgi/common: rename uart output file names and shell variables.
Rename uart output file name (the file used to capture the uart output)
and associated shell variables to match with the current uart
configuration.

TF-A boot, runtime and MM logs would be routed via secure port. The
file name and shell variable associated with this port are prefixed
accordingly. Likewise, uefi and OS logs would be routed via non-secure.
The file name and shell variable associated with this port are also
prefixed accordingly.

Signed-off-by: Rohit Mathew <rohit.mathew@arm.com>
Change-Id: Ibbf2599153620c92759942f9b7e40012d1250232