sgi575/run_model: update the value of quantum FVP parameter
Update the quantum value from 500 to 480. This change is required to
support Trusted Firmware Test Framework on SGI-575 platform as quantum
value greater than 480 makes the test framework hang in-between the test
cases. The quantum parameter is one of the FVP parameters that is used
to adjust the timing precision of the FVP.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: Iad1c924cadc4b75c4f67ad10a9f1e2f1336847b6